Time: 5 ms (81.69%) | Memory: 23 MB (42.16%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -8,7 +8,7 @@ public:
             int mid = start + (end - start) / 2;
             long long total = 0;
             for (size_t i = 0; i < piles.size(); ++i) {
-                total += (piles[i] + mid - 1) / mid;
+                total += ((long long)piles[i] + mid - 1) / mid;
                 if (total > h) {
                     break;
                 }
